Move user reservations page to screens folder
